Move stories to the models module
