+++ modules/features/endofyear/src/test/kotlin/au/com/shiftyjelly/pocketcasts/endofyear/EndOfYearViewModelTest.kt
@@ -3,18 +3,19 @@ package au.com.shiftyjelly.pocketcasts.endofyear
 import app.cash.turbine.Turbine
 import app.cash.turbine.TurbineTestContext
 import app.cash.turbine.test
-import au.com.shiftyjelly.pocketcasts.endofyear.Story.CompletionRate
-import au.com.shiftyjelly.pocketcasts.endofyear.Story.Cover
-import au.com.shiftyjelly.pocketcasts.endofyear.Story.Ending
-import au.com.shiftyjelly.pocketcasts.endofyear.Story.LongestEpisode
-import au.com.shiftyjelly.pocketcasts.endofyear.Story.NumberOfShows
-import au.com.shiftyjelly.pocketcasts.endofyear.Story.PlusInterstitial
-import au.com.shiftyjelly.pocketcasts.endofyear.Story.Ratings
-import au.com.shiftyjelly.pocketcasts.endofyear.Story.TopShow
-import au.com.shiftyjelly.pocketcasts.endofyear.Story.TopShows
-import au.com.shiftyjelly.pocketcasts.endofyear.Story.TotalTime
-import au.com.shiftyjelly.pocketcasts.endofyear.Story.YearVsYear
 import au.com.shiftyjelly.pocketcasts.models.to.RatingStats
+import au.com.shiftyjelly.pocketcasts.models.to.Story
+import au.com.shiftyjelly.pocketcasts.models.to.Story.CompletionRate
+import au.com.shiftyjelly.pocketcasts.models.to.Story.Cover
+import au.com.shiftyjelly.pocketcasts.models.to.Story.Ending
+import au.com.shiftyjelly.pocketcasts.models.to.Story.LongestEpisode
+import au.com.shiftyjelly.pocketcasts.models.to.Story.NumberOfShows
+import au.com.shiftyjelly.pocketcasts.models.to.Story.PlusInterstitial
+import au.com.shiftyjelly.pocketcasts.models.to.Story.Ratings
+import au.com.shiftyjelly.pocketcasts.models.to.Story.TopShow
+import au.com.shiftyjelly.pocketcasts.models.to.Story.TopShows
+import au.com.shiftyjelly.pocketcasts.models.to.Story.TotalTime
+import au.com.shiftyjelly.pocketcasts.models.to.Story.YearVsYear
 import au.com.shiftyjelly.pocketcasts.models.to.TopPodcast
 import au.com.shiftyjelly.pocketcasts.models.type.SubscriptionTier
 import au.com.shiftyjelly.pocketcasts.repositories.endofyear.EndOfYearManager
